+++ CHANGELOG.md
@@ -5,6 +5,51 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.19.4] - 2025-02-17
+### Details
+#### Chore
+- Lock file maintenance
+
+## [0.19.3] - 2025-02-17
+### Details
+#### Chore
+- Update dependency pnpm to v10.4.1
+
+## [0.19.2] - 2025-02-14
+### Details
+#### Chore
+- Update dependency tuist to v4.42.0
+
+## [0.19.1] - 2025-02-14
+### Details
+#### Chore
+- Update dependency pnpm to v10
+
+## [0.19.0] - 2025-02-14
+### Details
+#### Feat
+- Add a text prompt component
+
+## [0.18.3] - 2025-02-13
+### Details
+#### Chore
+- Lock file maintenance
+
+## [0.18.2] - 2025-02-13
+### Details
+#### Chore
+- Update dependency tuist to v4.41.0
+
+## [0.18.1] - 2025-02-12
+### Details
+#### Chore
+- Update dependency node to v22.14.0
+
+## [0.18.0] - 2025-02-12
+### Details
+#### Feat
+- Support runtime options in the SingleChoicePrompt
+
 ## [0.17.2] - 2025-02-03
 ### Details
 #### Chore
@@ -412,6 +457,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update dependency reddavis/asynchrone to from: "0.22.0"
 - Rename the project
 
+[0.19.4]: https://github.com/tuist/XcodeGraph/compare/0.19.3..0.19.4
+[0.19.3]: https://github.com/tuist/XcodeGraph/compare/0.19.2..0.19.3
+[0.19.2]: https://github.com/tuist/XcodeGraph/compare/0.19.1..0.19.2
+[0.19.1]: https://github.com/tuist/XcodeGraph/compare/0.19.0..0.19.1
+[0.19.0]: https://github.com/tuist/XcodeGraph/compare/0.18.3..0.19.0
+[0.18.3]: https://github.com/tuist/XcodeGraph/compare/0.18.2..0.18.3
+[0.18.2]: https://github.com/tuist/XcodeGraph/compare/0.18.1..0.18.2
+[0.18.1]: https://github.com/tuist/XcodeGraph/compare/0.18.0..0.18.1
+[0.18.0]: https://github.com/tuist/XcodeGraph/compare/0.17.2..0.18.0
 [0.17.2]: https://github.com/tuist/XcodeGraph/compare/0.17.1..0.17.2
 [0.17.1]: https://github.com/tuist/XcodeGraph/compare/0.17.0..0.17.1
 [0.17.0]: https://github.com/tuist/XcodeGraph/compare/0.16.29..0.17.0